Core/General: added bunch of updates for 4.3.4
	* movement updates should be all ported and working (maybe i forgot something)
	* update packets should be ok now
	* power system
	* SMSG_NEW_WORLD structure, SMSG_INIT_CURRENCIES (fixes wowerror)
	* achievement packets
	* spell casting (also ported some things from Trinity - NYI!)

+++ src/server/game/Spells/Auras/SpellAuras.cpp
@@ -248,7 +248,7 @@ void AuraApplication::ClientUpdate(bool remove)
         }
     }
 
-    data << uint8(flags);
+    data << uint16(flags);
     data << uint8(aura->GetCasterLevel());
     data << uint8(aura->GetStackAmount() > 1 ? aura->GetStackAmount() : (aura->GetCharges()) ? aura->GetCharges() : 1);
